Added AbstractGoalViolations class and made GoalViolations and IntraBrokerGoalViolations to extend it
